requirements.txt
# Telebot requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
pyTelegramBotAPI>=4.11.0
telebot>=0.0.5
spotipy>=2.23.0
requests>=2.29.0
beautifulsoup4>=4.12.2
anthropic
google-generativeai
